refactor(github): split 1275-line monolith into github/{http,profile,star-gate,index}.ts
- http.ts: shared fetch utilities, token pool, GraphQL/REST helpers
- profile.ts: getProfileSummary, calculateDetailedStreaks, rank/trophy/badge logic
- star-gate.ts: checkStarStatus, verifyAndInjectStar, getRepoStarCount
- index.ts: barrel re-export (import path unchanged: @/lib/github)
- Collapse 4 near-duplicate recursive paginators into 2 generic
  paginateForward/paginateBackward helpers (~250 lines removed)
- Move achievement cache from in-process Map to Redis (Finding 6 combined)

Breaking: none — all public exports preserved at @/lib/github
